+++ services/madoc-ts/src/frontend/site/pages/loaders/root-loader.tsx
@@ -121,10 +121,6 @@ export const RootLoader: UniversalComponent<RootLoaderType> = createUniversalCom
         creator: () => [],
         name: 'getUserDetails',
       },
-      {
-        creator: () => [],
-        name: 'getAutomatedUsers',
-      },
     ],
     getKey: () => {
       return ['root-config', []];