+++ CHANGELOG.md
@@ -5,13 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/digirati-co-uk/madoc-platform/compare/v2.2.4...main)
+## [Unreleased](https://github.com/digirati-co-uk/madoc-platform/compare/v2.2.5...main)
 <!-- 
 ### Fixed
 ### Added
 ### Changed
 ### Removed
 -->
+
+## [v2.2.5](https://github.com/digirati-co-uk/madoc-platform/compare/v2.2.4...v2.2.5) - 22/05/2024
 
 ### Fixed
 - Project feedback not working for all users